+++ lib/data/domain/model/conversation.dart
